ai_itinerary_scorer/requirements.txt
# Core FastAPI web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Document parsing and OCR
PyPDF2==3.0.1            # For basic PDF text extraction
pdfminer.six==20231228    # For advanced PDF parsing (column/layout)
python-docx==1.1.2        # For DOCX document parsing
pytesseract==0.3.10       # For OCR extraction from scanned PDFs
Pillow==10.3.0            # Image support for pytesseract

# For optional PDF-to-image for OCR
PyMuPDF==1.24.4           # (aka fitz) for page rasterization

# NLP and ML
spacy==3.7.4              # For NER/entity extraction pipeline
scikit-learn==1.4.2       # For regression & helpers
sentence-transformers==2.7.0  # For SBERT embeddings (includes transformers/torch)
faiss-cpu==1.8.0          # For efficient similarity search (CPU only)
transformers==4.40.0      # For advanced models/sentiment

# Database/Utilities
sqlite-utils==3.36        # For SQLite DB utility
pandas==2.2.2             # For CSV/TSV and EDA
numpy==1.26.4             # Core array math

# Configuration & environment
python-dotenv==1.0.1

# (Optional) For type hints and static analysis
pydantic==2.7.3

# (Optional) Progress bars during training/database population
tqdm==4.66.4

# Torch should be installed automatically via sentence-transformers, but if not:
# torch==2.3.0

# ----
# Notes:
# All packages are chosen for CPU-only environments and verified to work under Python 3.8+.
# For GPU support, you would want to customize torch/faiss installations.
# If deploying to Alpine Linux containers, you may need to adapt some builds, esp. PyMuPDF.

# --- End requirements ---
